weather_api/requirements.txt
Django==4.2.7
djangorestframework>=3.16.0
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.32.4
django-redis==5.4.0
redis==5.0.1
celery==5.3.4